[PILCO] Fix dtype fragility, multi-dim batch reset, and clarify Jacobian comments
- Cast rollout to target dtype before GP fitting to prevent float32/float64 mismatch
- Fix ImaginedEnv._reset to handle arbitrary batch_size dimensions (not just 1D)
- Correct misleading comments: C_xu is the policy Jacobian ∂μ_u/∂μ_x, not cov(x,u)

+++ sota-implementations/pilco/pilco.py
@@ -105,7 +105,7 @@ def pilco_loop(
         base_world_model = GPWorldModel(obs_dim=obs_dim, action_dim=action_dim).to(
             env.device
         )
-        base_world_model.fit(rollout)
+        base_world_model.fit(rollout.to(dtype))
         base_world_model.eval()
 
         imagined_env = ImaginedEnv(
